requirements.txt
rpaframework>=7.4.2
requests>=2.22.0
